requirements.txt
azure-mgmt-compute
azure-common